# openapi.yaml
openapi: 3.0.3
info:
  title: TopicsPulse API
  description: >
    Local, on-demand semantic topic analysis over user support messages.
    One ingestion endpoint (POST /messages), three read endpoints
    (GET /topics, GET /users/{login}/topics, GET /messages/search).
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local docker-compose instance

paths:
  /health:
    get:
      summary: Health check
      operationId: health
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /messages:
    post:
      summary: Ingest one message
      description: >
        Stores the message immediately (fast INSERT). Embedding is computed
        asynchronously by a background worker — this call does not wait for it.
      operationId: postMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestMessageRequest"
            examples:
              default:
                value:
                  login: user123
                  user_id: 987654
                  text: "Не могу вывести деньги"
                  created_at: "2026-09-21T12:30:00Z"
                  source: game
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
        "400":
          $ref: "#/components/responses/BadRequest"

  /topics:
    get:
      summary: Overall topics for a fixed period (precomputed nightly)
      description: >
        Reads the latest completed nightly analysis run for the given
        period. Never computes anything on request; returns 404 if the
        nightly job hasn't produced a result for this period yet.
      operationId: getTopics
      parameters:
        - name: period
          in: query
          description: One of the periods configured via ANALYSIS_PERIODS (default 24h,7d,30d).
          schema:
            type: string
            enum: [24h, 7d, 30d]
            default: 7d
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopicsResponse"
        "404":
          description: No completed analysis run yet for this period
        "400":
          $ref: "#/components/responses/BadRequest"

  /users/{login}/topics:
    get:
      summary: Topics for one user, computed on demand
      description: >
        Computed synchronously at request time (single LLM call over the
        user's messages in range) — not precomputed, no clustering step.
      operationId: getUserTopics
      parameters:
        - name: login
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopicsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  /messages/search:
    get:
      summary: Full-text search over all messages
      description: >
        Postgres tsvector/tsquery search (russian config), falling back to
        pg_trgm similarity search when the FTS query returns no rows.
      operationId: searchMessages
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: login
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
            enum: [game, forum, telegram]
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  /internal/analyze:
    post:
      summary: Manually trigger the nightly analysis job
      description: >
        Runs the same pipeline as the nightly cron, synchronously, for all
        configured periods. Intended for local testing only (no auth).
      operationId: triggerAnalysis
      responses:
        "200":
          description: OK

components:
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string

  schemas:
    IngestMessageRequest:
      type: object
      required: [login, text, created_at, source]
      properties:
        login:
          type: string
        user_id:
          type: integer
          format: int64
          nullable: true
        text:
          type: string
        created_at:
          type: string
          format: date-time
        source:
          type: string
          enum: [game, forum, telegram]

    Topic:
      type: object
      properties:
        rank:
          type: integer
        name:
          type: string
        summary:
          type: string
        message_count:
          type: integer
        unique_users:
          type: integer
        sources:
          type: array
          items:
            type: string
        representative_messages:
          type: array
          items:
            type: string
        topic_score:
          type: number
          format: double

    TopicsResponse:
      type: object
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        topics:
          type: array
          items:
            $ref: "#/components/schemas/Topic"

    SearchResult:
      type: object
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        text:
          type: string
        created_at:
          type: string
          format: date-time
        source:
          type: string
        rank:
          type: number
          format: double

    SearchResponse:
      type: object
      properties:
        query:
          type: string
        total:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"